backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
openai==1.55.3
python-dotenv==1.0.1
pydantic==2.10.3
jinja2==3.1.4
python-multipart==0.0.20
